requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.50.0
streamlit>=1.40.0
plotly>=5.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
